+++ trunk/source/AsioExpress/Platform/GetClockCountWin.cpp
@@ -7,7 +7,7 @@
 
 namespace AsioExpress {
 
-unsigned long GetClockCount(void)
+unsigned int GetClockCount(void)
 {
   return GetTickCount();
 }
